Move db initialization to a separate module and db env vars to __init__ file
