+++ src/collector/database/models.py
@@ -18,25 +18,11 @@
 #   along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import logging
-import os
-from playhouse.pool import PooledMySQLDatabase
 from peewee import (Model, PrimaryKeyField, IntegerField, BigIntegerField, DoubleField,
                     DateField, CharField)
-import sys
-from time import sleep
+from src.collector.database.db import db
 
 logger = logging.getLogger(__name__)
-
-
-db = PooledMySQLDatabase(
-    os.getenv('MYSQL_DATABASE'),
-    user=os.getenv('MYSQL_USER'),
-    password=os.getenv('MYSQL_PASSWORD'),
-    host=os.getenv('MYSQL_HOST'),
-    port=int(os.getenv('MYSQL_PORT', 3306)),
-    max_connections=8,
-    stale_timeout=300,
-)
 
 
 class BaseModel(Model):
@@ -83,21 +69,3 @@ class DailyStatsRecord(BaseModel):
     gas_fees_total_gwei = DoubleField(default=0)
     gas_fees_total_eth = DoubleField(default=0)
     gas_fees_total_usd = DoubleField(default=0)
-
-
-def wait_for_db():
-    for _ in range(30):
-        try:
-            db.connect()
-            db.close()
-            logger.info('Successfully connected to the database.')
-            return
-        except Exception as e:
-            logger.exception(e)
-            logger.warning(
-                f'Database connection failed. Retrying in {5} seconds...'
-            )
-            sleep(5)
-
-    logger.error('Failed to connect to the database after multiple attempts.')
-    sys.exit(1)
